Mc uframe

+++ ooiservices/manage.py
@@ -12,7 +12,7 @@ if os.environ.get('FLASK_COVERAGE'):
     COV = coverage.coverage(branch=True,include=basedir + '/app/*')
     COV.start()
 from ooiservices.app import create_app, db
-from flask.ext.script import Manager, Shell, Server
+from flask.ext.script import Manager, Shell, Server, prompt_bool
 from flask.ext.migrate import Migrate, MigrateCommand
 
 
@@ -24,7 +24,7 @@ def make_shell_context():
     from ooiservices.app.models import User, UserScope, UserScopeLink, Array
     from ooiservices.app.models import PlatformDeployment, InstrumentDeployment, Stream, StreamParameter, Watch
     from ooiservices.app.models import OperatorEvent
-    from ooiservices.app.models import Platformname, Instrumentname
+    from ooiservices.app.models import Platformname, Instrumentname, Annotation
 
 
     ctx = {"app": app,
@@ -40,7 +40,8 @@ def make_shell_context():
            "OperatorEvent": OperatorEvent,
            "StreamParameter": StreamParameter,
            "Platformname": Platformname,
-           "Instrumentname": Instrumentname}
+           "Instrumentname": Instrumentname,
+           "Annotation": Annotation}
     return ctx
 
 @manager.command
@@ -71,8 +72,9 @@ def test(coverage=False):
     if retval.failures:
         sys.exit(1)
 
+@manager.option('-bl', '--bulkload', default=False)
 @manager.option('-p', '--password', required=True)
-def deploy(password):
+def deploy(password, bulkload):
     from flask.ext.migrate import upgrade
     from ooiservices.app.models import User, UserScope, UserScopeLink, Array
     from ooiservices.app.models import PlatformDeployment, InstrumentDeployment, Stream, StreamParameterLink
@@ -86,20 +88,24 @@ def deploy(password):
     psql('-c', 'create database ooiuitest;', '-U', 'postgres')
     psql('ooiuitest', '-c', 'create schema ooiui')
     psql('ooiuitest', '-c', 'create extension postgis')
-    with open('db/ooiui_schema_before_data.sql') as f:
-        psql('ooiuidev', _in=f)
-    with open('db/ooiui_schema_data.sql') as f:
-        psql('ooiuidev', _in=f)
-    with open('db/ooiui_schema_after_data.sql') as f:
-        psql('ooiuidev', _in=f)
     db.create_all()
-    #os.system("psql ooiuidev < db/ooiui_schema_data.sql")
+    if bulkload:
+        with open('db/ooiui_schema_data.sql') as f:
+            psql('ooiuidev', _in=f)
+        app.logger.info('Bulk test data loaded.')
+
     # migrate database to latest revision
     #upgrade()
     app.logger.info('Insert default user, name: admin')
     User.insert_user(password)
     UserScope.insert_scopes()
 
+@manager.command
+def load_data():
+    from sh import psql
+    with open('db/ooiui_schema_data.sql') as f:
+        psql('ooiuidev', _in=f)
+    app.logger.info('Bulk test data loaded.')
 
 @manager.command
 def profile(length=25, profile_dir=None):
@@ -114,11 +120,15 @@ def profile(length=25, profile_dir=None):
 def destroy():
     from sh import psql
     db_check = str(db.engine)
-    if (db_check == 'Engine(postgres://postgres@localhost/ooiuidev)'):
-        psql('-c', 'drop database ooiuidev')
-        psql('-c', 'drop database ooiuitest')
-    else:
-        print 'Must be working on LOCAL_DEVELOPMENT to destroy db'
+    if prompt_bool(
+        "Are you sure you want to do drop %s" % db_check
+    ):
+        if (db_check == 'Engine(postgres://postgres@localhost/ooiuidev)'):
+            psql('-c', 'drop database ooiuidev')
+            psql('-c', 'drop database ooiuitest')
+            app.logger.info('ooiuidev and ooiuitest databases have been dropped.')
+        else:
+            print 'Must be working on LOCAL_DEVELOPMENT to destroy db'
 
 if __name__ == '__main__':
     manager.run()
